backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
flask-limiter==2.4.0
redis==3.5.3
python-dotenv==0.19.0
bcrypt==3.2.0
flask-migrate==3.1.0
flask-swagger-ui==3.36.0
python-dateutil==2.8.2
werkzeug==2.0.1
gunicorn==20.1.0
requests==2.28.1
